+++ src/demo_tezos_domains/handlers/on_storage_diff.py
@@ -14,17 +14,19 @@ async def on_storage_diff(ctx: HandlerContext, storage: NameRegistryStorage) -> 
             return
 
         if item.level == "1":
-            await models.TLD.update_or_create(id=record_name, defaults=dict(owner=item.owner))
+            await models.TLD.update_or_create(id=record_name, defaults={'owner': item.owner})
         else:
             if item.level == "2":
                 await models.Domain.update_or_create(
                     id=record_name,
-                    defaults=dict(
-                        tld_id=record_path[-1],
-                        owner=item.owner,
-                        expiry=storage.store.expiry_map.get(item.expiry_key) if item.expiry_key else None,  # type: ignore
-                        token_id=int(item.tzip12_token_id) if item.tzip12_token_id else None,
-                    ),
+                    defaults={
+                        'tld_id': record_path[-1],
+                        'owner': item.owner,
+                        'expiry': storage.store.expiry_map.get(item.expiry_key) if item.expiry_key else None,  # type: ignore
+                        'token_id': int(item.tzip12_token_id) if item.tzip12_token_id else None,
+                    },
                 )
 
-            await models.Record.update_or_create(id=record_name, defaults=dict(domain_id='.'.join(record_path[-2:]), address=item.address))
+            await models.Record.update_or_create(
+                id=record_name, defaults={'domain_id': '.'.join(record_path[-2:]), 'address': item.address}
+            )