+++ tests/integration_tests/test_demos.py
@@ -13,9 +13,11 @@ import demo_quipuswap.models
 import demo_tezos_domains.models
 import demo_tezos_domains_big_map.models
 import demo_tzcolors.models
+from dipdup.utils import skip_ci
 from dipdup.utils.database import tortoise_wrapper
 
 
+@skip_ci
 class DemosTest(IsolatedAsyncioTestCase):
     # TODO: store cache in xdg_cache_home, keep databases and logs after last run
     def setUp(self):
@@ -84,8 +86,8 @@ class DemosTest(IsolatedAsyncioTestCase):
             tlds = await demo_tezos_domains.models.TLD.filter().count()
             domains = await demo_tezos_domains.models.Domain.filter().count()
 
-            self.assertEqual(5, tlds)
-            self.assertEqual(237, domains)
+            self.assertEqual(1, tlds)
+            self.assertEqual(145, domains)
 
     async def test_tezos_domains_big_map(self):
         self.run_dipdup('tezos_domains_big_map.yml')
@@ -94,5 +96,5 @@ class DemosTest(IsolatedAsyncioTestCase):
             tlds = await demo_tezos_domains_big_map.models.TLD.filter().count()
             domains = await demo_tezos_domains_big_map.models.Domain.filter().count()
 
-            self.assertEqual(5, tlds)
-            self.assertEqual(237, domains)
+            self.assertEqual(1, tlds)
+            self.assertEqual(145, domains)